+++ src/index.ts
@@ -1,24 +1,16 @@
 // components exports
 export * from './components/blocks/UCodeBlock.js';
 export * from './components/blocks/UFileBlock.js';
-export * from './components/blocks/UJsonBlock.js';
 export * from './components/blocks/UMarkedBlock.js';
 export * from './components/blocks/UTextBlock.js';
-export * from './components/blocks/UThinkBlock.js';
-export * from './components/blocks/UToolBlock.js';
 export * from './components/blocks/URefBlock.js';
 export * from './components/blocks/UTableBlock.js';
-export * from './components/buttons/UAttachButton.js';
 export * from './components/buttons/UCopyButton.js';
-export * from './components/buttons/UVoteButton.js';
 export * from './components/message/UMessage.js';
 export * from './components/prompt/UPrompt.js';
 export * from './components/references/URefTag.js';
 export * from './components/references/URefCard.js';
 export * from './components/references/URefCardGroup.js';
-
-// intent exports
-export * from './components/intents/UQuestionIntent.js';
 
 // view exports
 export * from './components/views/UImagesView.js';
@@ -28,10 +20,8 @@ export * from './components/views/UView.js';
 
 // types exports
 export type * from './types/BlockItem';
-export type * from './types/JsonNode';
 export type * from './types/JsonSchema';
 export type * from './types/References';
-export * from './types/Intents';
 export * from './types/Views';
 
 // event types exports
@@ -39,6 +29,5 @@ export type * from './events/SendEvent';
 export type * from './events/StopEvent';
 
 // utilities exports
-export * from './utilities/IntentPromptBuilder.js';
 export * from './utilities/ViewPromptBuilder.js';
 // export * from './utilities/dom-agent/index.js';